Clamp force
The clamping is consistent with the python version, but the logic is
wrong. We should rescale the force vector to clamp the norm of the force
rather than clamping the invididual components of the vector. The way we
do it here, the clamping can change the direction of the vector and the
norm of the vector is not the expected one.