Set the state signal transmitter to None
The observer thread does not use the state broadcast signal. Tis means
that state_rx is not moved into the thread an is therefore dropped too
early. This leads to the state broadcaster failing to send its signals.

Here we create the state broadcaster with the signal tx set to None so
it does not try to send the signals. This change needsto be reverted if
the observer thread starts using signal_rx.