Reduce the components to the ones used by OpenMM
The other components are likely to be provided as CONECT records in the
PDB. This drastically reduces the size of the data to embed and makes
tests easier. It is also likely to reduce spurious bonds in MD
simulations as these systems tend to have custom residue names that can
clash with the PDB component list.