Use all 4 characters of the atom names in components
In the components crate, the tom names were encoded on only 3 characters
instead of 4. This meant that 4-character-long names where truncated so
the corresponding bonds were not added to the topologies.

In addition, shorter names were pruned making more missing bonds.

In this commit, we encode the atom names on 4 characters, and we pad
shorter names with null characters.