Allow shorter resnames in components
While the components crate assumed residue names to be 3 characters or
shorter, it was actually allowing only residue names that were exactly 3
characters long. This is an issue with nucleic acids, for instance, that
have residue names 1 or 2 characters long.

This commit prevents the python script that generates the components
from discarding residues with short names and pads the residue names
with null characters.