Implement frame resetting
Resetting the frame_index to 0 indicates to the client that the
accumulated FrameData should be reset with the current frame.

With this change, we modify the file format used to record trajectories:
instead of saving the FrameData, we save the GetFrameResponse as the
frame index becomes significant.

We also add a magic number and a file version number to the recorded
files.