Fix simulation lists with a single element
When the simulation list contains a single element, then the path
reduction function returns an empty string. Indeed, the common parts of
all the path are all the parts of that single path.

Instead, when there is only one element in the list of simulation, we
return the file name in most cases, and the full path in degenerate
cases where there is no file name.