[GUI] Remove the client when we remove the server
When the server fails to connect, the server gets removed but not the
client. Ehen we then connect with a different server (changing a used
port for a free one, for instance) we have a mismatch between the server
and the client so client requestget stuck and the UI freezes.

Here, we clear the client when the server errors to avoid the issue.
Stopping the server with the Stop button was already clearing the
client.