Start moving openmm specific logic in its own module
The definition of a `Simulation` and an `OpenMMSimulation` were
intertwined which made it difficult to identify what is generic to
simulations, regardless of engine, and what is specific to OpenMM,
currently the only implementation.

This commit only moves some definitions in separate files. Many concepts
are still intertwined, especially around computing user interactions.