Send the user forces in the FrameData
In order to compute the work associated with user interactions, we need
to access the forces due to these interactions. In this commit, the
forces are sent as a flatten list of coordinates under the key
`forces.user`. This may be fairly wasteful as most of the array is 0
even when there is an interaction going on.

This commit also refactors some of the book keeping in
simulation_thread.rs. This refactoring avoid cases where variables
related to a simulation could be out of sync.