Send the user forces as a sparse matrix
Sending the full matrix of user forces with one force per atom in the
simulation means doubling the size of the FrameData we send every frame.
Knowing that most of the force array contains 0 even with many
interactions, and that the size of a FrameData is limited by the way we
use gRPC, sending this full array is wasteful of bandwidth and message
size budget.

In this commit, we send the user forces as a sparse matrix with one
FrameData array containing the indices of the atoms that are touched and
an other array containing the forces for these atoms.