More simulation_thread refactoring
Move the cached user interactions and the frame count inside the
TrackedTrajectory. These variables were not reset when a new simulation
was loaded so they could be at odd with the actual state of the
simulation.

The commit also normalise some integer types to usize, which reduce the
need to convert between units in various places and gives a less verbose
code.