Simplify user energies in simulation_thread
The user energies were stored in the TrackedSimulation (and previously
in the thread closure) as `(Option<String>, Option<f64>)`. The idea
being that the energy for a given interaction would be stored alongside
the interaction ID. However, the ID is optional for a user interaction
so it may not be available; samewise a user interaction may not have an
energy as only the force is mandatory.

In practice, we only need the energies so we can report the total so we
do not need the interaction ID, and we do not need the interactions that
do not have an energy value. Now, the TrackedSimulation only stores the
summed energy.