Include velocities and forces in the FrameData
Add the --include-velocities and --include-forces flags to the command
line, and the corresponding check boxes in the GUI. These flags let a
server include the velocities and forces for all atoms in the FrameData
so it can be used by a client or recorded to file.

A visual client could use the data to colour particles or display force
vectors. More interestigly, the data can be used in analyses later on.

Fixes #185