Remove local path from narupa-proto
narupa-proto was refering to files on my disk but out of the project. These files were the proto file
for the protobuf well known types (e.g. struct and value) in my conda environemnt. I do not need to refer to these
files. Instead, I should install them alongside protoc. On ubuntu, this is done by installing the `protobuf-dev`
package; on fedore, the `protobuf-devel` package.
This is now documented for fedora in the README.