Vendor openmm-sys
This will make it easier to update and modify the OpenMM binding to fit
our specific needs. In particular, the original version of the bindings
are stuck with OpenMM 7.4.1 that is very outdated, and does not build
with a recent version of cmake.

The vendored version is copied from the commit
624d6206db64385b05b505894f09204b69816d66 of the original repository
found here: <https://github.com/uob-positron-imaging-centre/openmm-rust>

Note that the original repository used a mirror of the openmm git
repository for its submodule. Here, we use the official github
repository.